some smalls fixes

+++ models/carlaDatasetSimple.py
@@ -128,7 +128,7 @@ class CarlaDatasetSimple(Dataset):
         tl = torch.tensor([1, 0] if data['tl_state'] == 'Green' else [0, 1], dtype=torch.float16)
         v_aff = torch.tensor([data['lane_distance'], data['lane_orientation']]).float()
         sum_pds = (s == 0).sum()
-        pds = torch.tensor([1, 0] if sum_pds > 0 else [0, 1], dtype=torch.float16)
+        pds = torch.tensor([1, 0] if sum_pds > 0 else [0, 1]).float()
 
         return x, s, tl, v_aff, pds
